+++ package.json
@@ -9,10 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@types/node": "20.3.2",
-    "@types/react": "18.2.14",
-    "@types/react-dom": "18.2.6",
-    "@types/remark-prism": "^1.3.4",
     "autoprefixer": "^10.4.14",
     "eslint": "8.43.0",
     "eslint-config-next": "13.4.7",
@@ -20,15 +16,18 @@
     "postcss": "8.4.24",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "rehype-highlight": "^6.0.0",
+    "rehype-prism-plus": "^1.6.1",
     "rehype-stringify": "^9.0.3",
     "remark": "^14.0.3",
-    "remark-html": "^15.0.2",
     "remark-rehype": "^10.1.0",
     "tailwindcss": "^3.3.2",
     "typescript": "5.1.6"
   },
   "devDependencies": {
-    "@svgr/webpack": "^8.0.1"
+    "@svgr/webpack": "^8.0.1",
+    "@types/node": "20.3.2",
+    "@types/react": "18.2.14",
+    "@types/react-dom": "18.2.6",
+    "@types/remark-prism": "^1.3.4"
   }
 }
