+++ src/app/Header.tsx
@@ -16,13 +16,9 @@ const Header = () => {
     // fixed 요소는 스크롤에도 반응하지 않아 header를 fixed position으로 설정해야 함.
     return (
         <>
-            <div className="fixed top-0 h-10 w-full flex justify-center items-center bg-bg-primary z-50">
+            <div className="sticky top-0 h-10 w-full flex justify-center items-center bg-bg-primary z-50">
                 <button className="absolute left-4" onClick={() => handleClick()}>back</button>
                 <h1 className="text-lg">퀴즈</h1>
-            </div>
-
-            <div className="relative top-0 h-10 w-full z-0">
-
             </div>
         </>
