+++ src/app/quiz/types.ts
@@ -1,0 +1,8 @@
+export type QuizItem = {
+    item_content: string
+};
+
+export type Quiz = {
+    content: string,
+    items: QuizItem[]
+};
